+++ src/main/typescript/ruleEngine/ruleElements/flatModifier.ts
@@ -22,6 +22,10 @@ export default class FlatModifier extends RuleElement {
     this.apply();
   }
 
+  protected override _onAfterComputation(): void {
+    this.apply();
+  }
+
   /** Apply the rule element to the target Document. */
   protected apply(): void {
     if (typeof this.value !== "number") return;